Move `run_command()` to `run` module
