+++ src/shim.rs
@@ -16,18 +16,7 @@ pub fn python_shim(
     run(cfg, settings, command, arguments)
 }
 
-pub fn run_command(cfg: &Option<Cfg>, settings: &Settings, command_and_args: &str) -> Result<()> {
-    let s = shlex::split(&command_and_args)
-        .ok_or_else(|| format_err!("Failed to split command from {:?}", command_and_args))?;
-    let (cmd, arguments) = s.split_at(1);
-    let cmd = cmd
-        .get(0)
-        .ok_or_else(|| format_err!("Failed to extract command from {:?}", command_and_args))?;
-
-    run(cfg, settings, cmd, arguments)
-}
-
-fn run<S>(cfg: &Option<Cfg>, settings: &Settings, command: &str, arguments: &[S]) -> Result<()>
+pub fn run<S>(cfg: &Option<Cfg>, settings: &Settings, command: &str, arguments: &[S]) -> Result<()>
 where
     S: AsRef<str> + std::convert::AsRef<std::ffi::OsStr> + std::fmt::Debug,
 {
